+++ openid_provider/conf.py
@@ -12,3 +12,6 @@ FILESTORE_PATH = getattr(settings, 'OPENID_PROVIDER_FILESTORE_PATH',
 
 SREG_DATA_CALLBACK = getattr(settings, 'OPENID_SREG_DATA_CALLBACK',
                              'openid_provder.utils.get_default_sreg_data')
+
+AX_DATA_CALLBACK = getattr(setting, 'OPENID_AX_DATA_CALLBACK',
+                           'openid_provder.utils.get_default_ax_data')