+++ openid_provider/utils.py
@@ -3,7 +3,7 @@
 # some code from http://www.djangosnippets.org/snippets/310/ by simon
 # and from examples/djopenid from python-openid-2.2.4
 from openid_provider import conf
-from openid.extensions import sreg
+from openid.extensions import ax, sreg
 
 from django.core.exceptions import ImproperlyConfigured
 from django.utils.importlib import import_module
@@ -19,6 +19,15 @@ def get_default_sreg_data(request, orequest):
         'fullname': request.user.get_full_name(),
     }
 
+def get_default_ax_data(request, orequest):
+    return {
+        'http://axschema.org/contact/email': request.user.email,
+        'http://axschema.org/namePerson': request.user.get_full_name(),
+        'http://axschema.org/namePerson/friendly': request.user.username,
+        'http://axschema.org/namePerson/first': request.user.first_name,
+        'http://axschema.org/namePerson/last': request.user.last_name,
+    }
+
 def add_sreg_data(request, orequest, oresponse):
     callback = get_sreg_callback()
     if callback is None or not callable(callback):
@@ -28,9 +37,26 @@ def add_sreg_data(request, orequest, oresponse):
     sreg_resp = sreg.SRegResponse.extractResponse(sreg_req, sreg_data)
     oresponse.addExtension(sreg_resp)
 
+def add_ax_data(request, orequest, oresponse):
+    callback = get_ax_callback()
+    if callback is None or not callable(callback):
+        return
+    ax_data = callback(request, orequest)
+    ax_req = ax.FetchRequest.fromOpenIDRequest(orequest)
+    ax_resp = ax.FetchResponse(ax_req)
+    for attr in ax_req.getRequiredAttrs():
+        ax_resp.addValue(attr, ax_data.get(attr))
+    oresponse.addExtension(sreg_resp)
+
 def get_sreg_callback():
     try:
         return import_module_attr(conf.SREG_DATA_CALLBACK)
+    except (ImportError, AttributeError):
+        return None
+
+def get_ax_callback():
+    try:
+        return import_module_attr(conf.AX_DATA_CALLBACK)
     except (ImportError, AttributeError):
         return None